Perform manual postDraw signal cleanup in Codename adapter
This commit simply makes it so that the Codename adapter removes its
postDraw signal before switching states. This primarily prevents a
crash that could occur when exiting PlayState that occured from within
the postDraw callback of the adapter, during which it would attempt to
access the non-existent strumlines.

This also acts as a small performance optimization, where the signal is
removed when it's not necessary.